+++ packages/terra-search-field/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed port resized 320x256 px for `terra-search-field`.
+
 ## 3.99.0 - (January 9, 2024)
 
 * Changed